Migration build error resolved (#44)
* file change : api

* migrate ambigious directory name request into client

* Add pkg directory for containing lower bound dependancy-free packages (#42)

Merge pkg structure migration to Control logic.

None of functional changes has applied but, structural refactoring for reduce of dependency  has been applied.

* resolve previous build error

---------
