+++ api/start_vm.go
@@ -1,0 +1,30 @@
+package api
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/easy-cloud-Knet/KWS_Control/service"
+	"github.com/easy-cloud-Knet/KWS_Control/structure"
+)
+
+type ApiStartVmRequest struct {
+	UUID structure.UUID `json:"uuid"`
+}
+
+func (c *handlerContext) startVm(w http.ResponseWriter, r *http.Request) {
+	var req ApiStartVmRequest
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		http.Error(w, "Invalid request body", http.StatusBadRequest)
+		return
+	}
+	defer r.Body.Close()
+
+	err := service.StartVM(req.UUID, c.context)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(http.StatusOK)
+}
