+++ client/model/vm.go
