+++ service/vm.go
@@ -10,8 +10,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/easy-cloud-Knet/KWS_Control/request"
-	"github.com/easy-cloud-Knet/KWS_Control/request/model"
+	"github.com/easy-cloud-Knet/KWS_Control/client"
+	"github.com/easy-cloud-Knet/KWS_Control/client/model"
+	"github.com/easy-cloud-Knet/KWS_Control/pkg/guacamole"
+	internalssh "github.com/easy-cloud-Knet/KWS_Control/pkg/ssh"
 	"github.com/easy-cloud-Knet/KWS_Control/util"
 	"github.com/redis/go-redis/v9"
 
@@ -62,6 +64,7 @@ func CreateVM(w http.ResponseWriter, r *http.Request, contextStruct *vms.Control
 	aliveCount := 0
 
 	for i := range contextStruct.Cores {
+
 		core := &contextStruct.Cores[i]
 		log.DebugInfo("core %s checking: FreeMemory=%d, FreeCPU=%d, FreeDisk=%d, IsAlive=%t", core.IP, core.FreeMemory, core.FreeCPU, core.FreeDisk, core.IsAlive)
 
@@ -118,7 +121,7 @@ func CreateVM(w http.ResponseWriter, r *http.Request, contextStruct *vms.Control
 
 		return errors.New("selectedCore == nil")
 	}
-	var privateKeyPEM, publicKeyOpenSSH, err = GenerateSshKey()
+	var privateKeyPEM, publicKeyOpenSSH, err = internalssh.GenerateSSHKey()
 	if err != nil {
 		log.Error("GenerateSshKey() failed: %v", err, true)
 		return err
@@ -136,7 +139,7 @@ func CreateVM(w http.ResponseWriter, r *http.Request, contextStruct *vms.Control
 	cleanup := func() {
 		if guacamoleConfigured {
 			log.Info("clean up clean up")
-			if cleanupErr := CleanupGuacamoleConfig(string(uuid), contextStruct.GuacDB); cleanupErr != nil {
+			if cleanupErr := guacamole.Cleanup(string(uuid), contextStruct.GuacDB); cleanupErr != nil {
 				log.Error("Failed to cleanup Guacamole config during rollback: %v", cleanupErr)
 			}
 		}
@@ -169,7 +172,7 @@ func CreateVM(w http.ResponseWriter, r *http.Request, contextStruct *vms.Control
 	fmt.Printf("%s\n", cmsResp.MacAddr)
 	fmt.Printf("%s\n", cmsResp.SdnUUID)
 
-	userPass := GuacamoleConfig(req.Users[0].Name, string(req.UUID), cmsResp.IP, privateKeyPEM, contextStruct.GuacDB)
+	userPass := guacamole.Configure(req.Users[0].Name, string(req.UUID), cmsResp.IP, privateKeyPEM, contextStruct.GuacDB)
 
 	if userPass == "" {
 		log.Error("Failed to configure Guacamole", true)
@@ -222,8 +225,8 @@ func CreateVM(w http.ResponseWriter, r *http.Request, contextStruct *vms.Control
 	}
 
 	// Redis 저장 완료 후 HTTP 전송 (Core에서 Redis 업데이트 가능)
-	client := request.NewCoreClient(selectedCore)
-	_, err = client.CreateVM(context.Background(), req)
+	coreClient := client.NewCoreClient(selectedCore)
+	_, err = coreClient.CreateVM(context.Background(), req)
 	if err != nil {
 		log.Error("Error creating VM on core %s: %v", selectedCore.IP, err, true)
 		cleanup() // 직접 지우지 말고 요 함수 하나로--
@@ -263,8 +266,8 @@ func DeleteVM(uuid vms.UUID, contextStruct *vms.ControlContext, rdb *redis.Clien
 		return fmt.Errorf("VM with UUID %s not found", string(uuid))
 	}
 
-	client := request.NewCoreClient(core)
-	_, err := client.DeleteVM(context.Background(), model.DeleteVMRequest{
+	coreClient := client.NewCoreClient(core)
+	_, err := coreClient.DeleteVM(context.Background(), model.DeleteVMRequest{
 		UUID: uuid,
 		Type: model.HardDelete,
 	})
@@ -278,7 +281,7 @@ func DeleteVM(uuid vms.UUID, contextStruct *vms.ControlContext, rdb *redis.Clien
 		log.Error("error deleting instance %s from ControlContext: %v", uuid, err)
 		return err
 	}
-	if cleanupErr := CleanupGuacamoleConfig(string(uuid), contextStruct.GuacDB); cleanupErr != nil {
+	if cleanupErr := guacamole.Cleanup(string(uuid), contextStruct.GuacDB); cleanupErr != nil {
 		log.Error("Failed to cleanup Guacamole config during rollback: %v", cleanupErr)
 	}
 
@@ -298,8 +301,8 @@ func StartVM(uuid vms.UUID, contextStruct *vms.ControlContext) error {
 		return fmt.Errorf("VM with UUID %s not found", string(uuid))
 	}
 
-	client := request.NewCoreClient(core)
-	_, err := client.StartVM(context.Background(), model.StartVMRequest{
+	coreClient := client.NewCoreClient(core)
+	_, err := coreClient.StartVM(context.Background(), model.StartVMRequest{
 		UUID: uuid,
 	})
 	if err != nil {
@@ -316,8 +319,8 @@ func ShutdownVM(uuid vms.UUID, contextStruct *vms.ControlContext, rdb *redis.Cli
 		return fmt.Errorf("VM with UUID %s not found", string(uuid))
 	}
 
-	client := request.NewCoreClient(core)
-	_, err := client.ForceShutdownVM(context.Background(), model.ForceShutdownVMRequest{
+	coreClient := client.NewCoreClient(core)
+	_, err := coreClient.ForceShutdownVM(context.Background(), model.ForceShutdownVMRequest{
 		UUID: uuid,
 	})
 
@@ -355,9 +358,9 @@ func GetVMCpuInfo(uuid vms.UUID, contextStruct *vms.ControlContext) (model.CoreM
 		return model.CoreMachineCpuInfoResponse{}, errors.New(msg)
 	}
 
-	client := request.NewCoreClient(core)
+	coreClient := client.NewCoreClient(core)
 
-	cpuInfo, err := client.GetVMCpuInfo(context.Background(), uuid)
+	cpuInfo, err := coreClient.GetVMCpuInfo(context.Background(), uuid)
 	if err != nil {
 		msg := fmt.Sprintf("Error getting CPU info for VM %s on core %s: %v", uuid, core.IP, err)
 		log.Error(msg, true)
@@ -378,9 +381,9 @@ func GetVMMemoryInfo(uuid vms.UUID, contextStruct *vms.ControlContext) (model.Co
 		return model.CoreMachineMemoryInfoResponse{}, errors.New(msg)
 	}
 
-	client := request.NewCoreClient(core)
+	coreClient := client.NewCoreClient(core)
 
-	memoryInfo, err := client.GetVMMemoryInfo(context.Background(), uuid)
+	memoryInfo, err := coreClient.GetVMMemoryInfo(context.Background(), uuid)
 	if err != nil {
 		msg := fmt.Sprintf("Error getting memory info for VM %s on core %s: %v", uuid, core.IP, err)
 		log.Error(msg, true)
@@ -401,9 +404,9 @@ func GetVMDiskInfo(uuid vms.UUID, contextStruct *vms.ControlContext) (model.Core
 		return model.CoreMachineDiskInfoResponse{}, errors.New(msg)
 	}
 
-	client := request.NewCoreClient(core)
+	coreClient := client.NewCoreClient(core)
 
-	diskInfo, err := client.GetVMDiskInfo(context.Background(), uuid)
+	diskInfo, err := coreClient.GetVMDiskInfo(context.Background(), uuid)
 	if err != nil {
 		msg := fmt.Sprintf("Error getting disk info for VM %s on core %s: %v", uuid, core.IP, err)
 		log.Error(msg, true)
